Complete "1st styles version" for the "main-flow"
Complete the styles for the main flow of the project (select a pet and schedule a visit), the styles needs adjusts in class names and aspects like scalability with the use of components about BEM method.

Also remember that this is the first version of the styles for the project, in the next weeks it will be developed.

Other note is that the styles for searcher and info page must be nested in the "home-style.css" this is pendant also.

Other major pendant is that all the pages are designed for mobile, the only feature for desktop is the "max width" of 700px, is necessary to design the project for desktop