Styles for desktop version
I add the images for home page, backgrounds of hero section and info section. Then apply max width for all the pages, for a better view in desktop version, the width is setting on 700px. The background images also have a media query to change according the screen size, the min width is set on 500px. These are changes on the style and nothing in the JS code, only in CSS and some HTML.